readme: image cannot be pulled from the docker hub anymore.
Due to concerns raised by wowza, I have decided to remove the prebuilt
docker image from the docker hub. Users would have to build the image
locally if the wish to use this image to run wowza inside a docker
container.

I am not sure if this would be the end of the issue. In the worst case,
the project would be removed from github.com pending response from wowza.

Lets hope for the best.